Changes to Fabfile to support prism
A variety of changes to the fabfile to support running on the new
prism infrastructure.

Most are in the docker-machine -> openstack interface.

Details:

Changes in fab up:

- New controller URL on prism
- there appears to be a bug in docker-machine where the tenant-name command always makes a call
to the v2 api (rather than the v3 that prism is running) which causes a failure.
directly setting the tenant ID doesn't have this problem.

- image name format changed
- different flavors available
- new required parameter --openstack-domain-id default

- Remove .aws reqirements as we are not currently supporting ceph

- No longer need to remake the /mnt drive

- Remove the ceph reference storage location as it's no longer available