update fabfile - fix prism docker problems
"Fab configure" has stopped working on prism.
Suggests that something about the default configuration of the vm's spun up by
docker-machine has changed.

These changes in the def configure() allow it to run without error.

- removed commented-out code leftover from POD as it is no longer necessary

- added apt command to remove docker-ce-cli and containerd.io.
without removing docker-ce-cli , there is a conflict when we try to install docker-engine
from dpkg and it fails with error message: "docker-ce-cli conflicts with docker-engine"
containerd.io is removed as it is listed here when removing above packages:

The following packages were automatically installed and are no longer required:
 containerd.io docker-ce-cli libltdl7

(It turns out libltdl7 is a dependency for the docker we're about to install, so leave that in.)

Then, we service docker stop before installing docker-engine as otherwise it appears to be still
running and to crash when we try to install the engine:
"A dependency job for docker.service failed. See 'journalctl -xe' for details."

And start it up again after the new version is installed.