[#22] dont download RNAseq bam at all - move out of way
We don't want sequence data to be downloaded to downstream/ even temporarily;
instead, move the bam (on the VM) to parent dir so it's out of the way.
Improvement on this would be to give it a temp file name to avoid collision but
there's nothing else running that would make a sorted.bam directly in /mnt/outputs anyhow.