New options if fab up crashes
We've been observing a crash when 'fab up' runs.
docker-machine creates the openstack vm but when it tries to
install docker on it, it fails with:

E: Unable to locate package docker-compose-plugin

the workaround is to use --engine-install-url  to call the older
version of the install docker script that doesnt have that package.

so that may prevent the crash. but in case it doesn't, there's now
an alternate flow -- instead of running fab configure
you can run fab unlock installdocker

to get it into a good state where you can now fab reference , fab process.