removed grep -v -- not working.
so the previous version is broken because i forgot the pipe  character
but i tried putting it in - so the last line is
| grep -v "DEBUG toil"

and it's not sucessfully filtering the lines.

I'm not sure if the pipe is running inside or outside the docker
and im not sure whether docker is sending things to stdout or stderr or what.
So for now I just totally remove it.

(so no, there is not a committed version with the pipe in --
I tried running it without committing and it did run but didn't filter the lines. )