Fix fusion potential hang
(this change applies to both standard and ERCC-transcript runs)

fix situation where fusion would hang indefinitely if it didn't generate proper output and instead left behind a _STARtmp folder with a named pipe inside it -- fab would try to download the pipe and it would never say it was done.

With this version -- if it doesn't find any fusion output files at all, it will accept that and continue on with variants and jfkm before moving to the next sample.

this is the version of the fabfile i am testing right now